relocatable.bbclass: workaround for failure with empty pkgconfig dir.
This is a temporary workaround for build failures of some packages,
for which the native class fails to build. The failing task is
do_populate_sysroot in relocatable_native_pcfile. The actual failure
happens because ${pkgconfigdir} gets created without any actual .pc-
files being put there, which relocatable.bbclass cannot handle
gracefully.

Currently the only package known to fail is glib-networking. This patch
changes relocatable.bbclass to cope with the situation. It is a temporary
workaround until we manage to trace down the root cause of failure in
glib-networking (or some of its depndencies, the current suspicion
pointing to glib-2.0/gio).

Signed-off-by: Krisztian Litkey <kli@iki.fi>